[select] Refactor index tracking
